+++ src/definitions.hpp
@@ -1,6 +1,12 @@
 #ifndef DEFINITIONS_H
 #define DEFINITIONS_H
-#include "tracelog.h"
+
+#if defined(USE_TRACELOG)
+    #include "tracelog.h"
+    #define TRACELOG(level, ...) TraceLog(level, __VA_ARGS__)
+#else
+    #define TRACELOG(level, ...) (void)0
+#endif
 
 /**********************
 global